fix: keep provider plans on rendered lines
